Various changes
ContactNumber is now a bigint from an int
Server responses now demarcate each field with "::," and end with "::\n"
InsertNewDelivery now sends to the client the autoincrement that it assigned to the new Delivery; this should fix client's problems of not assigning any new units or accessories to the new Delivery
Added a prototype for a better threadserver.py

+++ rmonitor-data/pcrental.sql
@@ -47,7 +47,7 @@ CREATE TABLE Client (
 	Name varchar(100),
 	Address varchar(255),
 	ContactPerson varchar(30),
-	ContactNumber int
+	ContactNumber bigint
 );
 
 CREATE TABLE Parts (
